filebench: fix coredump problem with dirwidth=1
If we set dirwidth=1 when defining fileset, run
'filebench -f <wml-test>.f' will cause coredump.
Because we set fileset->fs_meandepth in fileset_populate() as follows,
$ fileset->fs_meandepth=log(entries+leafdirs)/log(meandirwidth).
where meandirwidth is equal to 1 as same with dirwidth in <wml-test>.f.
So fileset->fs_meandepth is set to inf, which will cause endless
recursion of fileset_populate_subdir(). Finally, coredump occurs.

Here, we will use a little bias (0.1) instead of log(1) when
meandirwidth is equal to 1.

Signed-off-by: Zhiqiang Liu <liuzhiqiang26@huawei.com>